app screen components moved to separate folder
