+++ src/screens/Game/Game.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Story, Meta } from "@storybook/react";
-import { Game, IGameProps } from "@/components/Game/Game";
+import { Game, IGameProps } from "@/screens/Game";
 import { Provider } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
 import { authSlice } from "@/auth";
