+++ src/screens/RegistrationScreen/index.ts
